Create tests for local hosts database
The hosts format is not very complex so we can have a handful of tests
that should cover almost everything.

We don't want to match partial host names: only the ones that fully
match will be looked up in that database.

Since this is the first test we also need to add pytest and pytest-mock
(used in the tests) to the requirements.txt file.