Move the lookup code to a more generic function
This opens up the door to running DNS lookups or any type or RR since
it's basically always the same procedure.

Usage of the dnspython library for performing the DNS lookups is
something that dns_deep_state should hide from its users: if we expose
objects specific to that library, we'll lock our users into the
underlying implementation. For now dnspython works well but we never
know what the future holds for us.