Start roughing out the main interface
Now that we have some basic tools and before venturing much further into
implementing how things are done, we should move out to designing the
main interface and some rough pseudocode for how the module will be
used.

This shall help us have a better direction for the whole thing and have
a more unified internal interface.

The docstrings are super verbose for now, and that'll probably go away
as things get more precise, but for now that's the idea that we'll be
following.