use setup.cfg for requirements instead of requirements.txt
usage of the requirements.txt file is an added manual step, e.g. running
`pip install -e .` installs nothing and you need to manually install the
requirements. That's super annoying and not very well integrated.

So to ease this bit up, let's use setup.cfg for all requirements.

This means that if one only wants to install the runtime requirements,
they can use:

    pip install .

and if one is prepping their dev environment in a venv, they can use:

    pip install -e .[test]

This change also neatly splits install vs dev requirements, which was
not clear before.