change return value of dns.full_report to a dictionary
it's a bit ridiculous to serialize intermediary reports to json to
immediately deserialize them in the level above.

this change makes the full_report methods on the individual probers less
self-sufficient, but for now that's all good. if those methods need to
output json at some point, we can add a param that disables
serialization