API change: move all of the reports into the same class
instead of calling down to the probes to have them report, we'll
concentrate all reporting into one class that directly uses the probe
methods to obtain what's necessary. This should keep the probe classes
more focused on obtaining information and we'll leave the aggregation
and structuring of the report data to the reporting class.

now that we're testing the hosts report in the context of the reporting
class, we're starting to see issues with checking that the hosts probe
does what we're expecting it to do: the other probes can also call
random things during instantiation. since we don't want to mock out
everything that the other probes are doing to test the one we're after,
we have to start "disabling" the other probes by mocking out their
classes.

also, since we're reusing the fixture for the hosts test, we'll import
it from that test module. to do that we need the tests directory to
become a package so that we can perform a relative import.