implement domain registry information query with RDAP
right now this new method is really just a dumb wrapper to the function
in the whoisit module. That's because the module seems to be doing
exactly what we want, but also we'll be adding some weight to this
information fetch method eventually to grab data from whois whenever
it's necessary.

For now hopefully, RDAP covers a good number of domain names already.