mock out unnecessary probes during report tests
we're now starting to see a bit more of the previously encountered
effect: whenever we're instantiating a report class, all probes also get
instantiated and thus they add some unwanted external calls.

with the new RDAP probe, it means that we're bootstrapping the RDAP
library on every report test! that has to go away for speed and sanity.