RDAP: fallback to a query that permits weak ssl if needed
Some RDAP servers currently are using old and weak algorithms in their
TLS communication. rdap.nominet.uk is one such case. We'd still like to
grab the information when that happens. The information that we're
grabbing is public information so there isn't anything too sensitive
that's transmitted. It's just using annoyingly weak crypto.

We don't want to enable weak crypto in normal cases, so we will only do
it specifically for the case of an ssl error of DH_KEY_TOO_SMALL.

The whoisit library author added this fallback when I asked for it, so
we'll need to use at least version 2.2.0 of the library.