warn about the very unfinished state and add intentions for the tool
huh .. apparently this project got to first page on hacker news. I'm
pretty surprised by this given the still pretty embryonic state.

Let's at least warn about the state of things. I haven't documented the
project and it's mostly because everything is still moving a lot and I'm
a tiny bit uncertain if the interface is close to sedimentation or if
big overhauls are still to come.

Still, I'll add some notes about the intentions for this project to at
least give an idea to folks what to expect from this. If it makes some
folks want to help out, great. If it makes some folks want to keep an
eye out for a more finished thing, that's also great.