Replace use of .format by f-strings
This new python syntax is way more elegant than calls to .format()

The specified minimum python version was already 3.6+ so with this
requirement, f-strings were already available.

If there's ever a request to support older versions than python 3.6,
we'll see how this can be achieved, but for now let's use more modern
features.