Implement SOA record lookup and lookup on a specific nameserver
With this we can now make sure that all nameservers are serving up the
same zone file.

I've realized that the expectations for the return value were wrong in
the test for the report so I changed that up. I had decided that the
DnsProbe would avoid letting library details surface up, so we'll
instead send a dictionary back.

The SOA lookup also needs to be performed on a specific nameserver, so
this functionality also needed to be implemented.