Set proper type annotation for the soa structure
pyright was complaining that it couldn't infer the type of values used
in the dictionary built as return value for dns_report()

This type definition is correct for now. We might need to change it as
the dns report gets populated with more information, but we'll see when
we get there.