Specify code package name instead of src base directory
From what I could find, the value given to the `--cov` parameter should
be a python module. `src` is not a python module, it's merely just a
directroy containing the main module. So let's try changing that param
and see if the coverage reports make more sense during CI runs.