test_soa is wrongly giving a host name instead of IP address
The dnspython library expects an IP address, not a host name. So if we
ever want to query a certain name server, users will have to resolve the
host name to an IP first.

The parameter names are not clear enough on the type of value expected,
so let's change that.