mock out lookup() instead of resolve() to decouple tests
now that we have a test for the successful calls of lookup(), we can
decouple all of the tests for specific rr types so that they won't fail
if lookup() is broken -- the tests directly on lookup itself should fail
in those cases.

those functions are really just shims on top of lookup() so we're really
just interested in seeing that the returned data is formatted as
expected.